+++ core/timetrace.go
@@ -2,28 +2,33 @@ package core
 
 import (
 	"errors"
-	"fmt"
+	"io"
+	"os"
 	"time"
 
 	"github.com/dominikbraun/timetrace/config"
 )
 
 var (
-	ErrNoEndTime          = errors.New("no end time for last record")
-	ErrTrackingNotStarted = errors.New("start tracking first")
+	ErrNoEndTime           = errors.New("no end time for last record")
+	ErrTrackingNotStarted  = errors.New("start tracking first")
+	ErrAllDirectoriesEmpty = errors.New("all directories empty")
 )
 
 type Report struct {
 	Current            *Record
 	TrackedTimeCurrent *time.Duration
 	TrackedTimeToday   time.Duration
+	BreakTimeToday     time.Duration
 }
 
 // Filesystem represents a filesystem used for storing and loading resources.
 type Filesystem interface {
 	ProjectFilepath(key string) string
+	ProjectBackupFilepath(key string) string
 	ProjectFilepaths() ([]string, error)
 	RecordFilepath(start time.Time) string
+	RecordBackupFilepath(start time.Time) string
 	RecordFilepaths(dir string, less func(a, b string) bool) ([]string, error)
 	RecordDirs() ([]string, error)
 	ReportDir() string
@@ -55,7 +60,7 @@ func New(config *config.Config, fs Filesystem) *Timetrace {
 // Since parallel work isn't supported, the previous work must be stopped first.
 func (t *Timetrace) Start(projectKey string, isBillable bool) error {
 	latestRecord, err := t.LoadLatestRecord()
-	if err != nil {
+	if err != nil && !errors.Is(err, ErrAllDirectoriesEmpty) {
 		return err
 	}
 
@@ -108,8 +113,14 @@ func (t *Timetrace) Status() (*Report, error) {
 		return nil, err
 	}
 
+	breakTimeToday, err := t.breakTime(now)
+	if err != nil {
+		return nil, err
+	}
+
 	report := &Report{
 		TrackedTimeToday: trackedTimeToday,
+		BreakTimeToday:   breakTimeToday,
 	}
 
 	// If the latest record has been stopped, there is no active time tracking.
@@ -126,6 +137,21 @@ func (t *Timetrace) Status() (*Report, error) {
 	report.TrackedTimeCurrent = &trackedTimeCurrent
 
 	return report, nil
+}
+
+func (t *Timetrace) breakTime(date time.Time) (time.Duration, error) {
+	records, err := t.loadAllRecordsSortedAscending(date)
+	if err != nil {
+		return 0, err
+	}
+
+	// add up the time between records
+	var breakTime time.Duration
+	for i := 0; i < len(records)-1; i++ {
+		breakTime += records[i+1].Start.Sub(*records[i].End)
+	}
+
+	return breakTime, nil
 }
 
 // Stop stops the time tracking and marks the current record as ended.
@@ -216,31 +242,34 @@ func (t *Timetrace) trackedTime(date time.Time) (time.Duration, error) {
 	return trackedTime, nil
 }
 
-// FormatTodayTime returns the formated string of the total
-// time of today follwoing the format convention
-func (report *Report) FormatTodayTime() string {
-	return formatDuration(report.TrackedTimeToday)
-}
-
-// FormatCurrentTime returns the formated string of the current
-// report time follwoing the format convention
-func (report *Report) FormatCurrentTime() string {
-	return formatDuration(*report.TrackedTimeCurrent)
-}
-
-// formatDuration formats the passed duration into a string.
-// The format will be "8h 24min". If the duration is less then 60 secods
-// the format will be "0h 0min 12sec".
-func formatDuration(duration time.Duration) string {
-
-	hours := int64(duration.Hours()) % 60
-	minutes := int64(duration.Minutes()) % 60
-	secods := int64(duration.Seconds()) % 60
-
-	// as by convention if the duarion is < then 60 secods
-	// return "0h 0min Xsec"
-	if hours == 0 && minutes == 0 {
-		return fmt.Sprintf("0h 0min %dsec", secods)
+func (t *Timetrace) isDirEmpty(dir string) (bool, error) {
+	openedDir, err := os.Open(dir)
+	if err != nil {
+		return false, err
 	}
-	return fmt.Sprintf("%dh %dmin", hours, minutes)
+	defer openedDir.Close()
+
+	// Attempt to read 1 file's name, if it fails with
+	// EOF, directory is empty
+	_, err = openedDir.Readdirnames(1)
+	if err == io.EOF {
+		return true, nil
+	}
+
+	return false, err
+}
+
+func (t *Timetrace) latestNonEmptyDir(dirs []string) (string, error) {
+	for i := len(dirs) - 1; i >= 0; i-- {
+		isEmpty, err := t.isDirEmpty(dirs[i])
+		if err != nil {
+			return "", err
+		}
+
+		if !isEmpty {
+			return dirs[i], nil
+		}
+	}
+
+	return "", ErrAllDirectoriesEmpty
 }